Added the `AffinityMixin` CPU affinity wrapper (#6534)
Merged `enable_cpu_affinity()` function as a wrapper class AffinityMixin
for all DataLoader types.

---------
